refactor: ui components api
