+++ zellij-tile/src/ui_components/mod.rs
@@ -1,0 +1,12 @@
+mod table;
+mod text;
+mod nested_list;
+mod ribbon;
+
+pub use zellij_utils::plugin_api;
+pub use zellij_utils::prost::{self, *};
+
+pub use table::*;
+pub use text::*;
+pub use nested_list::*;
+pub use ribbon::*;
